+++ include/pfnet/constr_AC_LIN_FLOW_LIM.h
@@ -24,8 +24,6 @@ typedef struct Constr_AC_LIN_FLOW_LIM_Data Constr_AC_LIN_FLOW_LIM_Data;
 Constr* CONSTR_AC_LIN_FLOW_LIM_new(Net* net);
 void CONSTR_AC_LIN_FLOW_LIM_init(Constr* c);
 void CONSTR_AC_LIN_FLOW_LIM_count_step(Constr* c, Branch* br, int t);
-void CONSTR_AC_LIN_FLOW_LIM_allocate(Constr* c);
-void CONSTR_AC_LIN_FLOW_LIM_clear(Constr* c);
 void CONSTR_AC_LIN_FLOW_LIM_analyze_step(Constr* c, Branch* br, int t);
 void CONSTR_AC_LIN_FLOW_LIM_eval_step(Constr* c, Branch* br, int t, Vec* v, Vec* ve);
 void CONSTR_AC_LIN_FLOW_LIM_store_sens_step(Constr* c, Branch* br, int t, Vec* sA, Vec* sf, Vec* sGu, Vec* sGl);