+++ include/pfnet/constr_DCPF.h
@@ -18,8 +18,6 @@
 Constr* CONSTR_DCPF_new(Net* net);
 void CONSTR_DCPF_init(Constr* c);
 void CONSTR_DCPF_count_step(Constr* c, Branch* br, int t);
-void CONSTR_DCPF_allocate(Constr* c);
-void CONSTR_DCPF_clear(Constr* c);
 void CONSTR_DCPF_analyze_step(Constr* c, Branch* br, int t);
 void CONSTR_DCPF_eval_step(Constr* c, Branch* br, int t, Vec* v, Vec* ve);
 void CONSTR_DCPF_store_sens_step(Constr* c, Branch* br, int t, Vec* sA, Vec* sf, Vec* sGu, Vec* sGl);