+++ include/pfnet/constr_FIX.h
@@ -18,8 +18,6 @@
 Constr* CONSTR_FIX_new(Net* net);
 void CONSTR_FIX_init(Constr* c);
 void CONSTR_FIX_count_step(Constr* c, Branch* br, int t);
-void CONSTR_FIX_allocate(Constr* c);
-void CONSTR_FIX_clear(Constr* c);
 void CONSTR_FIX_analyze_step(Constr* c, Branch* br, int t);
 void CONSTR_FIX_eval_step(Constr* c, Branch* br, int t, Vec* v, Vec* ve);
 void CONSTR_FIX_store_sens_step(Constr* c, Branch* br, int t, Vec* sA, Vec* sf, Vec* sGu, Vec* sGl);