+++ include/pfnet/constr_GEN_RAMP.h
@@ -18,8 +18,6 @@
 Constr* CONSTR_GEN_RAMP_new(Net* net);
 void CONSTR_GEN_RAMP_init(Constr* c);
 void CONSTR_GEN_RAMP_count_step(Constr* c, Branch* br, int t);
-void CONSTR_GEN_RAMP_allocate(Constr* c);
-void CONSTR_GEN_RAMP_clear(Constr* c);
 void CONSTR_GEN_RAMP_analyze_step(Constr* c, Branch* br, int t);
 void CONSTR_GEN_RAMP_eval_step(Constr* c, Branch* br, int t, Vec* v, Vec* ve);
 void CONSTR_GEN_RAMP_store_sens_step(Constr* c, Branch* br, int t, Vec* sA, Vec* sf, Vec* sGu, Vec* sGl);