+++ include/pfnet/constr_LOAD_PF.h
@@ -18,8 +18,6 @@
 Constr* CONSTR_LOAD_PF_new(Net* net);
 void CONSTR_LOAD_PF_init(Constr* c);
 void CONSTR_LOAD_PF_count_step(Constr* c, Branch* br, int t);
-void CONSTR_LOAD_PF_allocate(Constr* c);
-void CONSTR_LOAD_PF_clear(Constr* c);
 void CONSTR_LOAD_PF_analyze_step(Constr* c, Branch* br, int t);
 void CONSTR_LOAD_PF_eval_step(Constr* c, Branch* br, int t, Vec* v, Vec* ve);
 void CONSTR_LOAD_PF_store_sens_step(Constr* c, Branch* br, int t, Vec* sA, Vec* sf, Vec* sGu, Vec* sGl);