+++ include/pfnet/constr_NBOUND.h
@@ -21,8 +21,6 @@
 Constr* CONSTR_NBOUND_new(Net* net);
 void CONSTR_NBOUND_init(Constr* c);
 void CONSTR_NBOUND_count_step(Constr* c, Branch* br, int t);
-void CONSTR_NBOUND_allocate(Constr* c);
-void CONSTR_NBOUND_clear(Constr* c);
 void CONSTR_NBOUND_analyze_step(Constr* c, Branch* br, int t);
 void CONSTR_NBOUND_eval_step(Constr* c, Branch* br, int t, Vec* v, Vec* ve);
 void CONSTR_NBOUND_store_sens_step(Constr* c, Branch* br, int t, Vec* sA, Vec* sf, Vec* sGu, Vec* sGl);