+++ include/pfnet/constr_PAR_GEN_P.h
@@ -3,7 +3,7 @@
  *
  * This file is part of PFNET.
  *
- * Copyright (c) 2015-2017, Tomas Tinoco De Rubira.
+ * Copyright (c) 2015, Tomas Tinoco De Rubira.
  *
  * PFNET is released under the BSD 2-clause license.
  */
@@ -20,8 +20,6 @@
 Constr* CONSTR_PAR_GEN_P_new(Net* net);
 void CONSTR_PAR_GEN_P_init(Constr* c);
 void CONSTR_PAR_GEN_P_count_step(Constr* c, Branch* br, int t);
-void CONSTR_PAR_GEN_P_allocate(Constr* c);
-void CONSTR_PAR_GEN_P_clear(Constr* c);
 void CONSTR_PAR_GEN_P_analyze_step(Constr* c, Branch* br, int t);
 void CONSTR_PAR_GEN_P_eval_step(Constr* c, Branch* br, int t, Vec* v, Vec* ve);
 void CONSTR_PAR_GEN_P_store_sens_step(Constr* c, Branch* br, int t, Vec* sA, Vec* sf, Vec* sGu, Vec* sGl);