+++ include/pfnet/constr_REG_GEN.h
@@ -22,8 +22,6 @@
 Constr* CONSTR_REG_GEN_new(Net* net);
 void CONSTR_REG_GEN_init(Constr* c);
 void CONSTR_REG_GEN_count_step(Constr* c, Branch* br, int t);
-void CONSTR_REG_GEN_allocate(Constr* c);
-void CONSTR_REG_GEN_clear(Constr* c);
 void CONSTR_REG_GEN_analyze_step(Constr* c, Branch* br, int t);
 void CONSTR_REG_GEN_eval_step(Constr* c, Branch* br, int t, Vec* v, Vec* ve);
 void CONSTR_REG_GEN_store_sens_step(Constr* c, Branch* br, int t, Vec* sA, Vec* sf, Vec* sGu, Vec* sGl);