+++ include/pfnet/func.h
@@ -78,7 +78,7 @@ void FUNC_clear_error(Func * f);
 BOOL FUNC_has_error(Func* f);
 char* FUNC_get_error_string(Func* f);
 void FUNC_set_error(Func* f, char* error_string);
-void FUNC_update_network(Func* f);
+void FUNC_update(Func* f);
 Net* FUNC_get_network(Func* f);
 void FUNC_set_func_init(Func* f, void (*func)(Func* f));
 void FUNC_set_func_count_step(Func* f, void (*func)(Func* f, Branch* br, int t));